Split parquet mod
