+++ arro3-io/src/parquet/mod.rs
@@ -1,0 +1,5 @@
+mod reader;
+mod writer;
+
+pub(crate) use reader::{read_parquet, read_parquet_async};
+pub(crate) use writer::write_parquet;
